% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_iec.R
\name{print.survivehr_iec}
\alias{print.survivehr_iec}
\title{Print Method for IEC Results}
\usage{
\method{print}{survivehr_iec}(x, ...)
}
\arguments{
\item{x}{Object of class "survivehr_iec"}

\item{...}{Additional arguments (ignored)}
}
\description{
Print Method for IEC Results
}
