% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_iec.R
\name{print.survivehr_iec_eval}
\alias{print.survivehr_iec_eval}
\title{Print Method for IEC Evaluation Results}
\usage{
\method{print}{survivehr_iec_eval}(x, ...)
}
\arguments{
\item{x}{Object of class "survivehr_iec_eval"}

\item{...}{Additional arguments (ignored)}
}
\description{
Print Method for IEC Evaluation Results
}
