% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{survivehr_load_model}
\alias{survivehr_load_model}
\title{Load a SurvivEHR model bundle}
\usage{
survivehr_load_model(path)
}
\arguments{
\item{path}{file path created by \code{survivehr_save_model()}.}
}
\value{
Named list (model bundle) identical in structure to the original
bundle returned by \code{survivehr_finetune()}.
}
\description{
Load a SurvivEHR model bundle
}
\examples{
\dontrun{
ft2 <- survivehr_load_model("my_model.pt")
}
}
