% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{survivehr_save_model}
\alias{survivehr_save_model}
\title{Save a SurvivEHR model bundle}
\usage{
survivehr_save_model(model_bundle, path)
}
\arguments{
\item{model_bundle}{object returned by training functions.}

\item{path}{file path ending in \code{.pt}.}
}
\value{
Invisibly returns \code{path}.
}
\description{
Save a SurvivEHR model bundle
}
\examples{
\dontrun{
tmp <- tempfile(fileext = ".pt")
survivehr_save_model(ft, tmp)
ft2 <- survivehr_load_model(tmp)
unlink(tmp)
}
}
